requirements.txt
Django==3.1.1
device_detector==0.10
numpy==1.20.1
pandas==1.2.2
pytz==2020.1
requests==2.24.0
boto3==1.15.1
botocore==1.18.1
tqdm==4.51.0
djangorestframework==3.12.2
